+++ client/CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Changed mobile layout of landing page ([#428](https://github.com/project-lux/lux-frontend/issues/428)).
+
 ### Removed
 
 ### Fixed